style: 랜딩페이지 스크롤 애니메이션 적용
